Resolves naming issue
