+++ src/main/java/cli_ttt/consoleui/ConsoleInput.java
@@ -1,4 +1,4 @@
-package ttt.consoleui;
+package cli_ttt.consoleui;
 
 import ttt.GameException;
 import ttt.Input;
