+++ src/main/java/cli_ttt/game/GameConstructor.java
@@ -1,7 +1,9 @@
-package ttt.game;
+package cli_ttt.game;
 
 import ttt.Player;
-import ttt.players.PlayerFactory;
+import ttt.game.Board;
+import ttt.game.GameEngine;
+import cli_ttt.players.PlayerFactory;
 
 import java.util.List;
