+++ src/main/java/cli_ttt/game/GameMenu.java
@@ -1,7 +1,8 @@
-package ttt.game;
+package cli_ttt.game;
 
 import ttt.Display;
 import ttt.Input;
+import ttt.game.GameEngine;
 
 import java.util.ArrayList;
 import java.util.Arrays;
