+++ src/test/java/cli_ttt/ConsoleBoardTest.java
@@ -1,8 +1,8 @@
-package ttt;
+package cli_ttt;
 
 import org.junit.Before;
 import org.junit.Test;
-import ttt.consoleui.ConsoleBoard;
+import cli_ttt.consoleui.ConsoleBoard;
 import ttt.game.Board;
 import ttt.game.Marks;
