+++ src/test/java/cli_ttt/GameConstructorTest.java
@@ -1,7 +1,7 @@
-package ttt;
+package cli_ttt;
 
+import cli_ttt.game.GameConstructor;
 import org.junit.Test;
-import ttt.game.GameConstructor;
 import ttt.game.GameEngine;
 
 import java.util.Arrays;
